+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Changed
 - (Breaking) Migrate Database and Memory Cleanup to be Active instead of Passive. Added in a config to control the threads Sleep timer.
+- (Breaking) throttle database updates based on threshold and sweep @santhosh7403
+santhosh7403
 
 ## 0.18.0 (2. February, 2026)