+++ CHANGELOG.md
@@ -5,10 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
+## 0.19.0 (17. March, 2026)
 ### Changed
 - (Breaking) Migrate Database and Memory Cleanup to be Active instead of Passive. Added in a config to control the threads Sleep timer.
 - (Breaking) throttle database updates based on threshold and sweep @santhosh7403
-santhosh7403
+- Added Database update upon Session memory clear to ensure database is synced if persistent.
 
 ## 0.18.0 (2. February, 2026)